requirements.txt
scikit-learn
torch
matplotlib
scipy
pandas
numpy
Pillow